refactor: #150 - checkRequestEligibility 테스트를 책임 단위로 분리
